refactor: extract GettingStartedChecklist component
Moves the inline renderGettingStartedChecklist render function from
App.jsx into src/components/GettingStartedChecklist.jsx. Theme variables
and callbacks (onDismiss, onComplete) are passed as props to all three
call sites.
