+++ src/App.jsx
@@ -23,6 +23,7 @@ import FrameEditor from './components/FrameEditor.jsx';
 import QuickAddFrameForm from './components/QuickAddFrameForm.jsx';
 import SmartSchedulePanel from './components/SmartSchedulePanel.jsx';
 import SuggestionAutocomplete from './components/SuggestionAutocomplete.jsx';
+import GettingStartedChecklist from './components/GettingStartedChecklist.jsx';
 
 // Encode a string that may contain non-ASCII characters as Base64.
 // btoa() throws InvalidCharacterError for codepoints > 255 (CJK, emoji, etc.).
@@ -10452,61 +10453,6 @@ const DayPlanner = () => {
   const allGettingStartedComplete = gettingStartedItems.every(item => item.completed);
   const gettingStartedCompleteCount = gettingStartedItems.filter(item => item.completed).length;
 
-  const renderGettingStartedChecklist = () => {
-    const completedCount = gettingStartedCompleteCount;
-    const totalCount = gettingStartedItems.length;
-    const progressPct = Math.round((completedCount / totalCount) * 100);
-    return (
-      <div className={`mb-4 rounded-lg border ${darkMode ? 'border-blue-500/30 bg-blue-500/10' : 'border-blue-200 bg-blue-50'} overflow-hidden`}>
-        <div className="px-3 py-2.5 flex items-center justify-between">
-          <div className="flex items-center gap-2">
-            <Sparkles size={15} className="text-blue-500" />
-            <span className="text-sm font-semibold text-blue-600 dark:text-blue-400">Getting Started</span>
-            <span className={`text-xs px-1.5 py-0.5 rounded-full ${darkMode ? 'bg-blue-500/30 text-blue-300' : 'bg-blue-200 text-blue-700'}`}>
-              {completedCount}/{totalCount}
-            </span>
-          </div>
-          <button
-            onClick={() => setGettingStartedDismissed(true)}
-            className={`${textSecondary} hover:${textPrimary} p-0.5`}
-            title="Dismiss"
-          >
-            <X size={14} />
-          </button>
-        </div>
-        {/* Progress bar */}
-        <div className={`mx-3 mb-2 h-1.5 rounded-full ${darkMode ? 'bg-gray-700' : 'bg-blue-100'}`}>
-          <div className="h-full rounded-full bg-blue-500 transition-all duration-300" style={{ width: `${progressPct}%` }} />
-        </div>
-        <div className="px-3 pb-3 space-y-1">
-          {gettingStartedItems.map(item => (
-            <div key={item.id} className={`flex items-center gap-2 py-0.5 text-sm ${item.completed ? (darkMode ? 'text-gray-500' : 'text-stone-400') : textSecondary}`}>
-              {item.completed
-                ? <CheckCircle size={14} className="text-blue-500 flex-shrink-0" />
-                : <div className={`w-3.5 h-3.5 rounded-full border ${darkMode ? 'border-gray-600' : 'border-stone-300'} flex-shrink-0`} />
-              }
-              <span className={item.completed ? 'line-through' : ''}>{item.label}</span>
-            </div>
-          ))}
-        </div>
-        <div className="px-3 pb-3">
-          <button
-            onClick={() => {
-              setTasks(prev => prev.filter(t => !t.isExample));
-              setUnscheduledTasks(prev => prev.filter(t => !t.isExample));
-              setRecycleBin(prev => prev.filter(t => !t.isExample));
-              setRecurringTasks(prev => prev.filter(t => !t.isExample));
-              setOnboardingComplete(true);
-              setGettingStartedDismissed(true);
-            }}
-            className={`w-full py-2 rounded-lg text-sm font-medium transition-colors ${darkMode ? 'bg-blue-500/20 text-blue-300 hover:bg-blue-500/30' : 'bg-blue-100 text-blue-700 hover:bg-blue-200'}`}
-          >
-            I'm Good to Go
-          </button>
-        </div>
-      </div>
-    );
-  };
 
   // Check if user has zero real tasks (for showing onboarding)
   const hasZeroRealTasks = useMemo(() => {
@@ -13835,7 +13781,22 @@ const DayPlanner = () => {
                   </div>
                 )}
                 {/* Getting Started checklist */}
-                {showGettingStarted && renderGettingStartedChecklist()}
+                {showGettingStarted && <GettingStartedChecklist
+                  items={gettingStartedItems}
+                  completedCount={gettingStartedCompleteCount}
+                  darkMode={darkMode}
+                  textPrimary={textPrimary}
+                  textSecondary={textSecondary}
+                  onDismiss={() => setGettingStartedDismissed(true)}
+                  onComplete={() => {
+                    setTasks(prev => prev.filter(t => !t.isExample));
+                    setUnscheduledTasks(prev => prev.filter(t => !t.isExample));
+                    setRecycleBin(prev => prev.filter(t => !t.isExample));
+                    setRecurringTasks(prev => prev.filter(t => !t.isExample));
+                    setOnboardingComplete(true);
+                    setGettingStartedDismissed(true);
+                  }}
+                />}
                 {/* Reschedule Tasks — shown when overdue past-day tasks exist, or incomplete today tasks after 7pm */}
                 {aiConfig?.enabled && aiConfig.features?.aiReschedule && gtdFrames.filter(f => f.enabled).length > 0 && (() => {
                   const _todayStr = getTodayStr();
@@ -17441,7 +17402,22 @@ const DayPlanner = () => {
                         </div>
                       )}
                       {/* Getting Started checklist */}
-                      {showGettingStarted && renderGettingStartedChecklist()}
+                      {showGettingStarted && <GettingStartedChecklist
+                  items={gettingStartedItems}
+                  completedCount={gettingStartedCompleteCount}
+                  darkMode={darkMode}
+                  textPrimary={textPrimary}
+                  textSecondary={textSecondary}
+                  onDismiss={() => setGettingStartedDismissed(true)}
+                  onComplete={() => {
+                    setTasks(prev => prev.filter(t => !t.isExample));
+                    setUnscheduledTasks(prev => prev.filter(t => !t.isExample));
+                    setRecycleBin(prev => prev.filter(t => !t.isExample));
+                    setRecurringTasks(prev => prev.filter(t => !t.isExample));
+                    setOnboardingComplete(true);
+                    setGettingStartedDismissed(true);
+                  }}
+                />}
                       {/* Reschedule Tasks — shown when overdue past-day tasks exist, or incomplete today tasks after 7pm */}
                       {aiConfig?.enabled && aiConfig.features?.aiReschedule && gtdFrames.filter(f => f.enabled).length > 0 && (() => {
                         const _todayStr = getTodayStr();
@@ -18625,7 +18601,22 @@ const DayPlanner = () => {
                     </div>
                   )}
                   {/* Getting Started checklist */}
-                  {showGettingStarted && renderGettingStartedChecklist()}
+                  {showGettingStarted && <GettingStartedChecklist
+                  items={gettingStartedItems}
+                  completedCount={gettingStartedCompleteCount}
+                  darkMode={darkMode}
+                  textPrimary={textPrimary}
+                  textSecondary={textSecondary}
+                  onDismiss={() => setGettingStartedDismissed(true)}
+                  onComplete={() => {
+                    setTasks(prev => prev.filter(t => !t.isExample));
+                    setUnscheduledTasks(prev => prev.filter(t => !t.isExample));
+                    setRecycleBin(prev => prev.filter(t => !t.isExample));
+                    setRecurringTasks(prev => prev.filter(t => !t.isExample));
+                    setOnboardingComplete(true);
+                    setGettingStartedDismissed(true);
+                  }}
+                />}
                   {/* Reschedule Tasks — shown when overdue past-day tasks exist, or incomplete today tasks after 7pm */}
                   {aiConfig?.enabled && aiConfig.features?.aiReschedule && gtdFrames.filter(f => f.enabled).length > 0 && (() => {
                     const _todayStr = getTodayStr();
